# openapi/brindle.v1.yaml
openapi: 3.1.0
info:
  title: Brindle
  version: '2026-06-15'
  summary: Transactional email and SMS over one HTTP API.
  description: |
    Brindle takes a message, renders it, hands it to a carrier and tells you what happened to it.

    Every request under `/v1` carries `Authorization: Bearer <key>`. Keys begin `bk_live_` or
    `bk_test_`; a test key is accepted and everything it does is recorded, and nothing leaves
    Brindle. Every answer carries `Brindle-Version`, which is the dated version of the API this
    deployment speaks, and the `X-RateLimit-*` headers for the account the key belongs to.

    Refusals are the same object everywhere: `type`, `title`, `status`, `detail` and `code`. Branch
    on `code`; `detail` is written for a person and may be reworded.
  contact:
    name: Brindle support
    email: support@brindle.dev
    url: https://docs.brindle.dev
servers:
  - url: https://api.brindle.dev
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/templates:
    post:
      summary: Create a template
      operationId: createTemplate
      tags: [Templates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplate'
            examples:
              email:
                value:
                  name: Service due
                  channel: email
                  subject: Your {{item}} is due a service
                  body: <p>Hello {{first_name}},</p>
      responses:
        '201':
          description: The template, at version 1.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateWithBody'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      summary: List templates
      operationId: listTemplates
      tags: [Templates]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of templates, newest first.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Template'
                  next_cursor:
                    $ref: '#/components/schemas/NextCursor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/templates/{id}:
    parameters:
      - $ref: '#/components/parameters/TemplateId'
    get:
      summary: Read one template
      operationId: getTemplate
      tags: [Templates]
      responses:
        '200':
          description: The template and the wording of its current version.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateWithBody'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    put:
      summary: Write a new version of a template
      operationId: updateTemplate
      tags: [Templates]
      description: |
        The wording is never edited in place. A `PUT` writes the next version and points the
        template at it; older versions stay readable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTemplate'
      responses:
        '200':
          description: The template, now at the new version.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateWithBody'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/templates/{id}/versions:
    parameters:
      - $ref: '#/components/parameters/TemplateId'
    get:
      summary: The versions of a template
      operationId: listTemplateVersions
      tags: [Templates]
      responses:
        '200':
          description: Every version, oldest first.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateVersion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/templates/{id}/render:
    parameters:
      - $ref: '#/components/parameters/TemplateId'
    post:
      summary: Render a template without sending it
      operationId: renderTemplate
      tags: [Templates]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                variables:
                  $ref: '#/components/schemas/Variables'
                version:
                  type: integer
                  minimum: 1
                  description: The version to render. The current one otherwise.
            examples:
              current:
                value:
                  variables:
                    first_name: Esme
                    order_number: W-4410
      responses:
        '200':
          description: The rendered wording.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                type: object
                required: [subject, body, template_id, version]
                properties:
                  subject:
                    type: [string, 'null']
                  body:
                    type: string
                  template_id:
                    type: string
                  version:
                    type: integer
              examples:
                rendered:
                  value:
                    subject: W-4410 is on its way
                    body: <p>Hi Esme,</p>
                    template_id: tmpl_13jprr39jzhs
                    version: 2
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/webhooks:
    post:
      summary: Register an endpoint
      operationId: createWebhook
      tags: [Webhooks]
      description: |
        The signing secret is in this answer and in no other. Keep it; endpoints are not readable
        with their secret again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                  description: An https endpoint.
                events:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/EventType'
            examples:
              outcomes:
                value:
                  url: https://api.wainstall-cycles.co.uk/hooks/brindle
                  events: [delivered, bounced, failed]
      responses:
        '201':
          description: The endpoint, with its secret.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookWithSecret'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      summary: List endpoints
      operationId: listWebhooks
      tags: [Webhooks]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of endpoints, newest first.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
                  next_cursor:
                    $ref: '#/components/schemas/NextCursor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/webhooks/{id}:
    parameters:
      - $ref: '#/components/parameters/WebhookId'
    patch:
      summary: Turn an endpoint on or off
      operationId: setWebhookActive
      tags: [Webhooks]
      description: |
        An endpoint that is off keeps its history; deliveries simply stop being written for it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [active]
              properties:
                active:
                  type: boolean
            examples:
              off:
                value:
                  active: false
      responses:
        '200':
          description: The endpoint as it now stands.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      summary: Remove an endpoint
      operationId: deleteWebhook
      tags: [Webhooks]
      responses:
        '204':
          description: The endpoint and its delivery records are gone.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer bk_live_...`. A `bk_test_` key is accepted everywhere a live key is,
        and nothing it sends leaves Brindle.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: The `next_cursor` from the page before this one.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
      description: How many rows to return. Twenty-five otherwise, one hundred at most.
...
    TemplateId:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: tmpl_13jprr39jzhs
    WebhookId:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: wh_0g389xqn9q
  headers:
    BrindleVersion:
      description: The dated version of the API this deployment speaks.
      schema:
        type: string
        examples: ['2026-06-15']
...
    RateLimitLimit:
      description: Requests this account may make in a window.
      schema:
        type: integer
    RateLimitRemaining:
      description: Requests left in the current window.
      schema:
        type: integer
    RateLimitReset:
      description: |
        The Unix second at which the current window ends. A window opens on the account's first
        request and runs for a minute.
      schema:
        type: integer
  responses:
    BadRequest:
      description: The request could not be read.
      headers:
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            notAnAddress:
              value:
                type: https://docs.brindle.dev/errors/invalid_request
                title: Invalid request
                status: 400
                detail: 'to: "wren at brookmail" is not an email address.'
                code: invalid_request
    Unauthorized:
      description: No key, or a key Brindle does not know.
      headers:
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            noKey:
              value:
                type: https://docs.brindle.dev/errors/authentication_required
                title: Authentication required
                status: 401
                detail: 'Send your key as an Authorization header: `Authorization: Bearer bk_live_...`.'
                code: authentication_required
    NotFound:
      description: There is no such object on this account.
      headers:
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            noTemplate:
              value:
                type: https://docs.brindle.dev/errors/template_not_found
                title: Template not found
                status: 404
                detail: There is no template tmpl_nosuchthing on this account.
                code: template_not_found
...
    Unprocessable:
      description: The request was read and cannot be carried out.
      headers:
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            suppressed:
              value:
                type: https://docs.brindle.dev/errors/recipient_suppressed
                title: Recipient suppressed
                status: 422
                detail: ada.wardle@brookmail.com is on your suppression list. Take it off the list if you mean to write to it again.
                code: recipient_suppressed
    PayloadTooLarge:
      description: The request body is larger than this API accepts.
      headers:
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            tooBig:
              value:
                type: https://docs.brindle.dev/errors/payload_too_large
                title: Payload too large
                status: 413
                detail: That request body is larger than this API accepts.
                code: payload_too_large
    RateLimited:
      description: The account has spent its requests for this window.
      headers:
        Retry-After:
          description: Seconds to wait.
          schema:
            type: integer
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Channel:
      type: string
      enum: [email, sms]
...
    EventType:
      type: string
      enum: [queued, sent, delivered, bounced, opened, clicked, complaint, failed]
...
    NextCursor:
      type: [string, 'null']
      description: Pass this back as `cursor` for the next page. Null on the last page.
    Variables:
      type: object
      additionalProperties:
        type: string
      description: A value for every `{{placeholder}}` in the template.
    Problem:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
          description: The stable name of the refusal. Branch on this.
...
    CreateTemplate:
      type: object
      required: [name, channel, body]
      properties:
        name:
          type: string
          maxLength: 120
        channel:
          $ref: '#/components/schemas/Channel'
        subject:
          type: string
          maxLength: 200
          description: Required for an email template, and refused on an SMS one.
        body:
          type: string
          maxLength: 100000
    UpdateTemplate:
      type: object
      required: [body]
      properties:
        name:
          type: string
          maxLength: 120
        subject:
          type: string
          maxLength: 200
          description: Required on every version of an email template, and refused on an SMS one.
        body:
          type: string
          maxLength: 100000
    Template:
      type: object
      required: [id, name, channel, current_version, created_at, updated_at]
      properties:
        id:
          type: string
          examples: [tmpl_13jprr39jzhs]
        name:
          type: string
        channel:
          $ref: '#/components/schemas/Channel'
        current_version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TemplateWithBody:
      allOf:
        - $ref: '#/components/schemas/Template'
        - type: object
          properties:
            subject:
              type: [string, 'null']
            body:
              type: string
    TemplateVersion:
      type: object
      required: [id, template_id, version, subject, body, created_at]
      properties:
        id:
          type: string
        template_id:
          type: string
        version:
          type: integer
        subject:
          type: [string, 'null']
        body:
          type: string
        created_at:
          type: string
          format: date-time
...
    Webhook:
      type: object
      required: [id, url, events, active, created_at]
      properties:
        id:
          type: string
          examples: [wh_0g389xqn9q]
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventType'
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
    WebhookWithSecret:
      allOf:
        - $ref: '#/components/schemas/Webhook'
        - type: object
          required: [secret]
          properties:
            secret:
              type: string
              description: The signing secret. It is in this answer only.
              examples: [whsec_9wxm11zn2gmr1qxbntspgg907f3w6pxz]
...